Streamlined for command line execution.
If it is packaged in a .jar file, the program can now take arguments
from the command line to automate "-a" with a defined limiter. Ex: $
java -jar MontyHallSim.jar -a 1000 will automate the program and run it
through 1000 trials. Furthermore the program will no longer print
feedback during automation, speeding up the process significantly. In
addition, a process time has been added if automated (this is not
reliable, but it gives a sense of speed).